refactor: melhorando o sistema de validações de preenchimento dos dados no cadastro e edição de usuários

+++ src/Components/Modal/Excluir/Assistidos.js
@@ -7,6 +7,9 @@ import Loader from "../../Loader";
 import { useExcluirAssistido } from "../../../Services/excluirAssistido";
 import ButtonExcluir from "./Items";
 
+//Validadores
+import { validarProfessor } from "../../../Utils/validadores";
+
 export default function Assistidos ({assistido}){
     
     const { excluirAssistido, excluindo } = useExcluirAssistido();
@@ -14,9 +17,13 @@ export default function Assistidos ({assistido}){
     //função para chamar a requisição de excluir assistido
     const handleDeleteAssistido = async (e, id_uuid) => {
         e.preventDefault();
+        
+        if (!validarProfessor(localStorage.getItem("role"))){
+            return
+        }
 
-        const data = {"id_uuid": id_uuid};
-        //console.log("selecionado o assistido de id: ", id_uuid, " e nome: ", assistido.name)
+        const data = {id_uuid};
+        //console.log(data)
         
         await excluirAssistido(data);
         window.location.reload();
@@ -24,7 +31,7 @@ export default function Assistidos ({assistido}){
 
     return (
         <>
-            {excluindo === true ? <Loader /> : null}
+            {excluindo ? <Loader /> : null}
             <form onSubmit={(e) => handleDeleteAssistido(e, assistido.id_uuid)} className="flex flex-col">
                 <p className="text-sm">
                     Você realmente deseja excluir o assistido <span className="font-semibold text-red">{assistido.name}</span>? Ao excluir o assistido você exclui todos os processos vinculados a ele.
